Add VM start/stop
Work around https://github.com/cypress-io/cypress/issues/2636: We can't
set a static baseUrl before the test starts (as the server is running,
and cypress complains), which would avoid restarting the test after the
first `cy.visit()`. So set it as soon as it is available, and live with
the test restarting and losing all state after that.

This requires the `startVM` task to be idempotent.